requirements.txt

fastapi
uvicorn[standard]
python-multipart
faster-whisper
sqlalchemy
pytest
pytest-asyncio
hypothesis
httpx
python-magic
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
# Diarization Dependencies
torch
torchaudio
soundfile
speechbrain
huggingface-hub<0.25
scikit-learn
numpy
requests
# Audio Enhancement
# (noisereduce & pydub removed for performance optimization)
# AI Analysis
sumy
# Security & Performance
slowapi
# Tier 2 Infrastructure
psycopg2-binary
redis